+++ CHANGELOG.md
@@ -6,6 +6,7 @@ https://keepachangelog.com/en/1.0.0/
 - `UnnecessaryEnumerableMaterialization`: now supports conditional access, i.e. `values?.ToArray().ToList()`
 - `ExplicitEnumValues` and `EnumWithoutDefaultValue`: Downgraded from warning to info to reduce their prevalence
 - `UnnecessaryEnumerableMaterialization` tests use `VerifyCS`
+- `LinqTraversalBeforeFilter`: Fixed typo in diagnostic message
 
 ## [1.21.0] - 2023-01-21
 - `StringConcatenatedInLoop`: A `string` was concatenated in a loop which introduces intermediate allocations. Consider using a `StringBuilder` or pre-allocated `string` instead.